:art: move view children on their own file
