+++ src/renderer/views/dashboard/Dashboard.tsx
@@ -1,7 +1,5 @@
 import React from "react";
 
-import { Card } from "../../components/common/card/Card";
-import { isMailMainInfos, usePstStore } from "../../store/PSTStore";
 import style from "./Dashboard.module.scss";
 import { DashboardActions } from "./DashboardActions";
 import { DashboardImpact } from "./DashboardImpact";
@@ -13,28 +11,6 @@ export interface DashboardComponentProps {
     className?: string;
 }
 
-export const DashboardMailBody: React.FC<DashboardComponentProps> = ({
-    className,
-}) => {
-    const { mainInfos } = usePstStore();
-
-    if (!mainInfos) return null; // TODO: loader
-
-    return (
-        <Card title="Mail" color="grey" className={className}>
-            <div className={style.dashboardMail}>
-                {isMailMainInfos(mainInfos) ? (
-                    <div style={{ overflow: "scroll" }}>
-                        {mainInfos.data.email.contentText}
-                    </div>
-                ) : (
-                    <div data-i18n="TODO">Empty</div>
-                )}
-            </div>
-        </Card>
-    );
-};
-
 export const Dashboard: React.FC = () => {
     return (
         <div className={style.dashboard}>
@@ -45,7 +21,6 @@ export const Dashboard: React.FC = () => {
                 <div className={style.dashboardInfos}>
                     <DashboardRecap />
                     <DashboardInformations />
-                    <DashboardMailBody />
                 </div>
             </div>
         </div>
